Move addAttendance mutation to be used by other components
